libobs-d3d11: DXGI display capture buffering
Add helper thread for each desktop duplicator in an attempt to capture
in-between frames that are skipped when sampling once per frame on the
graphics thread.

There are two SPSC queues involved, one from the helper thread to the
graphics thread to queue up frames to render, and one from the graphics
thread to the helper thread to recycle frames for reuse.